+++ packages/gamut/src/Form/_variables.ts
@@ -1,9 +1,9 @@
-import { colors } from '@codecademy/gamut-styles';
+import { theme } from '@codecademy/gamut-styles';
 
 const variables = {
-  borderColor: colors['gray-200'],
-  itemBackground: colors.white,
-  itemBackgroundHover: colors['gray-200'],
+  borderColor: theme.colors['gray-200'],
+  itemBackground: theme.colors.white,
+  itemBackgroundHover: theme.colors['gray-200'],
   borderRadius: 2,
   transitionTime: '0.15s',
   formPadding: '0.8rem',